# it-helpdesk-openapi.json
{"openapi":"3.1.0","info":{"title":"Enterprise IT Support Copilot","description":"Citation-grounded technical support for OpenShift/SNO, watsonx Orchestrate, IBM Bob, and registered IBM products such as Guardium, Instana, Verify, Cloud Pak for Data, and Cloud Pak for Integration. Answers use approved indexed content and bounded live retrieval from registered official IBM documentation sources.","version":"0.1.0"},"servers":[{"url":"https://left-appraiser-disorder.ngrok-free.dev"}],"paths":{"/v1/assist":{"post":{"tags":["assist"],"summary":"Submit a technical support question","description":"Submit a technical support question for OpenShift/SNO, watsonx Orchestrate, IBM Bob, or another registered IBM product such as Guardium, Instana, Verify, Cloud Pak for Data, or Cloud Pak for Integration, as well as broad IBM and watsonx product-portfolio questions. For other registered IBM products, use the ibm_products domain or omit requested_scope and allow the service to resolve it. The response status is authoritative: ANSWERED returns answer_markdown, NEEDS_CLARIFICATION returns clarification_question, and the remaining statuses describe insufficient evidence, out-of-scope, invalid, or service-error outcomes. answer_markdown contains a visible source banner, safe suggested next steps, and a deterministic Sources section with clickable URLs. source_urls exposes those links structurally, and retrieval_provenance reports which retrieval paths were attempted and which sources support the final answer.","operationId":"assist_v1_assist_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssistRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyAuth":[]}]}},"/v1/domains":{"get":{"tags":["assist"],"summary":"List indexed knowledge domains","description":"Returns the domains currently indexed in the knowledge base with their chunk counts. Does not consume watsonx.ai tokens.","operationId":"domains_v1_domains_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DomainsResponse"}}}}},"security":[{"APIKeyAuth":[]}]}},"/healthz":{"get":{"tags":["health"],"summary":"Healthz","description":"Liveness probe — no dependency checks.","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/readyz":{"get":{"tags":["health"],"summary":"Readyz","description":"Verify core dependencies and enabled retrieval fallbacks.","operationId":"readyz_readyz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadyResponse"}}}}}}}},"components":{"schemas":{"AssistRequest":{"properties":{"question":{"type":"string","maxLength":2000,"minLength":3,"title":"Question","description":"The user's technical question, passed without rewriting its meaning."},"conversation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Conversation Id"},"conversation_context":{"items":{"$ref":"#/components/schemas/ConversationMessage"},"type":"array","maxItems":4,"title":"Conversation Context"},"requested_scope":{"$ref":"#/components/schemas/RequestedScope","description":"Optional explicit routing hints. An empty object is valid; the backend can infer registered products and versions from the question."}},"type":"object","required":["question"],"title":"AssistRequest"},"AssistResponse":{"properties":{"request_id":{"type":"string","title":"Request Id"},"status":{"type":"string","enum":["ANSWERED","NEEDS_CLARIFICATION","INSUFFICIENT_EVIDENCE","OUT_OF_SCOPE","INVALID_REQUEST","ERROR"],"title":"Status"},"intent":{"anyOf":[{"type":"string","enum":["qa","troubleshoot","summarize","unsupported"]},{"type":"null"}],"title":"Intent"},"answer_markdown":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Answer Markdown"},"clarification_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Clarification Question"},"citations":{"items":{"$ref":"#/components/schemas/Citation"},"type":"array","title":"Citations"},"source_urls":{"items":{"type":"string"},"type":"array","title":"Source Urls","description":"Deduplicated clickable HTTP(S) URLs cited by the final answer."},"suggested_next_steps":{"items":{"type":"string"},"type":"array","title":"Suggested Next Steps","description":"Safe follow-up prompts that help the user continue the task."},"retrieval_provenance":{"$ref":"#/components/schemas/RetrievalProvenance","description":"Shows whether OpenSearch, bounded IBM Docs retrieval, cached pages, or an external internet-search provider was used."},"safety_note":{"type":"string","title":"Safety Note","default":"Guidance is based only on the approved knowledge base; verify commands in your environment."},"trace_id":{"type":"string","title":"Trace Id"}},"type":"object","required":["status"],"title":"AssistResponse"},"Citation":{"properties":{"citation_id":{"type":"string","title":"Citation Id"},"title":{"type":"string","title":"Title"},"product":{"type":"string","title":"Product"},"ocp_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ocp Version"},"product_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Product Version"},"page_start":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page Start"},"page_end":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Page End"},"section_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Section Path"},"document_id":{"type":"string","title":"Document Id"},"chunk_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Chunk Id"},"source_uri":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Uri"},"retrieval_source":{"type":"string","enum":["opensearch_index","metadata_catalog","ibm_docs_cache","live_ibm_docs","internet_search"],"title":"Retrieval Source","description":"Where the evidence for this citation was obtained.","default":"opensearch_index"}},"type":"object","required":["citation_id","title","product","document_id"],"title":"Citation"},"ConversationMessage":{"properties":{"role":{"type":"string","enum":["user","assistant"],"title":"Role"},"content":{"type":"string","maxLength":2000,"minLength":1,"title":"Content"}},"type":"object","required":["role","content"],"title":"ConversationMessage"},"DomainInfo":{"properties":{"domain_id":{"type":"string","title":"Domain Id"},"display_name":{"type":"string","title":"Display Name"},"chunk_count":{"type":"integer","title":"Chunk Count"}},"type":"object","required":["domain_id","display_name","chunk_count"],"title":"DomainInfo"},"DomainsResponse":{"properties":{"domains":{"items":{"$ref":"#/components/schemas/DomainInfo"},"type":"array","title":"Domains"}},"type":"object","required":["domains"],"title":"DomainsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","const":"ok","title":"Status"}},"type":"object","required":["status"],"title":"HealthResponse"},"ReadyResponse":{"properties":{"status":{"type":"string","enum":["ready","degraded"],"title":"Status"},"opensearch":{"type":"boolean","title":"Opensearch"},"watsonx":{"type":"boolean","title":"Watsonx"},"adaptive_retrieval":{"type":"boolean","title":"Adaptive Retrieval"},"live_ibm_docs_enabled":{"type":"boolean","title":"Live Ibm Docs Enabled"},"live_ibm_docs_configured":{"type":"boolean","title":"Live Ibm Docs Configured"},"internet_search_enabled":{"type":"boolean","title":"Internet Search Enabled"},"internet_search_configured":{"type":"boolean","title":"Internet Search Configured"}},"type":"object","required":["status","opensearch","watsonx","adaptive_retrieval","live_ibm_docs_enabled","live_ibm_docs_configured","internet_search_enabled","internet_search_configured"],"title":"ReadyResponse"},"RequestedScope":{"properties":{"domain_id":{"anyOf":[{"type":"string","enum":["ocp_sno_support","watsonx_orchestrate","ibm_bob","ibm_products"]},{"type":"null"}],"title":"Domain Id","description":"Optional retrieval domain. Use ocp_sno_support for OpenShift/SNO, watsonx_orchestrate for watsonx Orchestrate/ADK, ibm_bob for IBM Bob, and ibm_products for all other registered IBM products. Omit the field when uncertain and let the service infer the domain."},"ocp_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ocp Version"},"deployment_type":{"anyOf":[{"type":"string","enum":["SNO","standard","compact"]},{"type":"null"}],"title":"Deployment Type"},"component":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Component"},"product":{"anyOf":[{"type":"string","maxLength":200,"minLength":1},{"type":"null"}],"title":"Product","description":"Canonical product name when known, primarily for ibm_products. Do not invent a product name; omit it when uncertain."},"product_version":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Product Version","description":"Product version stated by the user. Preserve it exactly and omit it when the user did not specify a version."}},"type":"object","title":"RequestedScope"},"RetrievalProvenance":{"properties":{"answer_sources":{"items":{"type":"string","enum":["opensearch_index","metadata_catalog","ibm_docs_cache","live_ibm_docs","internet_search"]},"type":"array","title":"Answer Sources","description":"Knowledge sources actually cited in the final answer."},"opensearch_searched":{"type":"boolean","title":"Opensearch Searched","default":false},"metadata_catalog_searched":{"type":"boolean","title":"Metadata Catalog Searched","default":false},"live_ibm_docs_retrieved":{"type":"boolean","title":"Live Ibm Docs Retrieved","default":false},"internet_search_performed":{"type":"boolean","title":"Internet Search Performed","default":false},"internet_search_provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Internet Search Provider"},"summary":{"type":"string","title":"Summary","default":"No retrieval source was used."}},"type":"object","title":"RetrievalProvenance"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyAuth":{"type":"apiKey","description":"API key for the Enterprise IT Support Copilot service.","in":"header","name":"X-API-Key"}}}}